Agrego sweetAlert2

+++ src/context/LibrosContext.jsx
@@ -1,6 +1,7 @@
 import { createContext, useContext } from "react";
 import { useNavigate } from "react-router";
 import fetchHandler from "../../utils/fetch";
+import Swal from "sweetalert2";
 
 const LibrosContext = createContext()
 
@@ -10,21 +11,41 @@ const LibrosProvider = ({children})=>{
     const navigate = useNavigate()
 
     const handleDelete = (id)=>{
-        const fetchDeleteBook = async () =>{
-            try {
-                const resp = await fetchHandler('http://localhost:3000/books/' + id,{
-                    method: 'DELETE',
-                })
-                if(resp){
-                    console.log(resp)
+        const fetchDeleteBook = async () => { 
+        await Swal.fire({
+            title: '¿Estás seguro?',
+            text: "¡No podrás revertir esto!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#d33',
+            cancelButtonColor: '#3085d6',
+            confirmButtonText: 'Sí, borrarlo',
+            cancelButtonText: 'Cancelar'
+        }).then(async (result) => {
+            if (result.isConfirmed) {
+                try {
+                    const resp = await fetchHandler('http://localhost:3000/books/' + id, {
+                        method: 'DELETE',
+                    });
+                    if (resp) {
+                    await Swal.fire({
+                        title:'¡Borrado!',
+                        text:'El libro ha sido eliminado correctamente.',
+                        icon: 'success',
+                        confirmButtonColor: '#d33',
+                        confirmButtonText: 'Okey',
+                        
+                    });
+                    navigate('/')
+                    }
+                } catch (error) {
+                    Swal.fire('Error', 'No se pudo eliminar el libro.', 'error');
                 }
-            } catch (error) {
-                console.log(error)
-                //SweetAlert
             }
-        }
-        fetchDeleteBook()
-        navigate('/')
+        });
+    }
+    fetchDeleteBook()
+
     }
 
       const handleAgregarLibro = async (nuevoLibro) => {
